# Conflicts:
#	src/Assets/Prefabs/Player/OriginSubMarine/マテリアル.005.mat
#	src/Assets/Scenes/Stages/DebugStageHaru/DebugStageHaru.unity

+++ src/Assets/Scripts/GameMain/Task/Enemy1Task.cs
@@ -24,11 +24,13 @@ namespace GameMain.Task
     public class Enemy1Task : BaseTask
     {
         [Header("爆破ダメージ")] [SerializeField] private uint attackPoint;
+        [SerializeField] private uint attackPointToOtherTask;
         [SerializeField] private int workerDamageCount;
         [SerializeField] private float explodeStartWaitTime;
         [SerializeField] private float disableBodyDelay;
         [SerializeField] private LayerMask damageLayer;
         [SerializeField] private float damageRangeFixOffset;
+        [SerializeField] private bool showExplodeRange;
 
         [Header("ダメージとスケールの比例関数")]
         [SerializeField]
@@ -53,17 +55,19 @@ namespace GameMain.Task
         private RuntimeNavMeshBaker navMeshBaker;
         private Material bodyMaterial;
         private Collider[] workerDamageBuffer;
-        private List<Collider> workerDamageBufferList;
+        private List<Collider> damageBufferList;
         [SerializeField] private Color blinkColor;
 
         private bool isAdsorption;
         private Transform adsorptionTarget;
         private Tween blinkTween;
         private Vector3 adsorptionOffset;
+        private Vector3 initialScale;
         private static readonly int IsWalking = Animator.StringToHash("IsWalking");
         private static readonly int CanBomb = Animator.StringToHash("CanBomb");
         private static readonly int BodyColor = Shader.PropertyToID("_Color");
 
+        public event Action OnBomb;
 
         public override void Initialize(IObjectResolver container)
         {
@@ -71,14 +75,12 @@ namespace GameMain.Task
             playerStatus = container.Resolve<PlayerStatus>();
             navMeshBaker = container.Resolve<RuntimeNavMeshBaker>();
             bodyMaterial = bodyRenderer.material;
-            workerDamageBuffer = new Collider[64];
-            workerDamageBufferList = new List<Collider>();
+            workerDamageBuffer = new Collider[128];
+            damageBufferList = new List<Collider>();
+            initialScale = transform.localScale;
 
             decalProjector.enabled = false;
             simpleAgent.SetActive(false);
-
-            SetDetection(false);
-            base.Disable();
 
             OnProgressChanged += UpdateScale;
             OnProgressChanged += UpdateBlinkColor;
@@ -96,17 +98,7 @@ namespace GameMain.Task
 
         private void UpdateScale(float progress)
         {
-            scaleBody.localScale = Vector3.one * damageScaleCurve.Evaluate(progress);
-        }
-
-        private void OnEnable()
-        {
-            decalProjector.enabled = true;
-            simpleAgent.SetActive(true);
-
-            animator.SetBool(IsWalking, true);
-
-            navMeshBaker?.Bake().Forget();
+            scaleBody.localScale = initialScale * damageScaleCurve.Evaluate(progress);
         }
 
         private void Update()
@@ -162,6 +154,7 @@ namespace GameMain.Task
             DisableBody().Forget();
 
             //爆破エフェクト開始
+            OnBomb?.Invoke();
             await bombEffectEvent.Bomb();
 
             DetectExplosionArea();
@@ -185,54 +178,63 @@ namespace GameMain.Task
             float radius = transform.localScale.x * explodeEffectSphere.localScale.x * damageRangeFixOffset;
             int count = Physics.OverlapSphereNonAlloc(origin, radius, workerDamageBuffer, damageLayer);
 
-            workerDamageBufferList.AddRange(workerDamageBuffer.Take(count));
+            damageBufferList.AddRange(workerDamageBuffer.Take(count));
         }
 
         private void Damage()
         {
-            int damagedCount = 0;
-            bool isPlayerDamaged = false;
+            int workerDamagedCount = 0;
 
-            //プレイヤーにダメージを与える & ワーカーを倒すまで繰り返す
-            while (workerDamageBufferList.Count > 0 && (damagedCount < workerDamageCount || !isPlayerDamaged))
+            foreach (Transform obj in damageBufferList.Shuffle().Select(item => item.transform))
             {
-                int index = Random.Range(0, workerDamageBufferList.Count);
-                Transform obj = workerDamageBufferList[index].transform;
-
                 //Damage to player
                 if (obj.CompareTag("Player"))
                 {
                     playerStatus.RemoveHp(attackPoint);
-                    isPlayerDamaged = true;
-                }
-                //Damage to worker
-                else if (obj.parent != null && obj.parent.TryGetComponent(out Worker worker))
-                {
-                    if (!worker.IsLocked)
-                    {
-                        worker.Kill();
-                        damagedCount++;
-                    }
+                    continue;
                 }
 
-                workerDamageBufferList.RemoveAt(index);
+                //Damage to worker
+                if (workerDamagedCount < workerDamageCount && obj.parent != null && obj.parent.TryGetComponent(out Worker worker))
+                {
+                    if (worker.IsLocked)
+                        continue;
+
+                    worker.Kill();
+                    workerDamagedCount++;
+                }
+
+                //Damage to other tasks
+                if (obj.parent != null && obj.parent.TryGetComponent(out BaseTask baseTask))
+                {
+                    if (baseTask == this || !baseTask.AcceptAttacks)
+                        continue;
+
+                    baseTask.ForceWork((int)attackPointToOtherTask);
+                }
             }
 
-            workerDamageBufferList.Clear();
+            damageBufferList.Clear();
         }
 
 
         public void ForceEnable()
         {
-            gameObject.SetActive(true);
+            animator.SetBool(IsWalking, true);
+            decalProjector.enabled = true;
+            simpleAgent.SetActive(true);
+            navMeshBaker?.Bake().Forget();
         }
 
         public override void Enable() { }
 
         public override void Disable() { }
 
-        private void OnDrawGizmosSelected()
+        private void OnDrawGizmos()
         {
+            if (!showExplodeRange)
+                return;
+
             Vector3 origin = explodeEffectSphere.position;
             float radius = transform.localScale.x * explodeEffectSphere.localScale.x * damageRangeFixOffset;
             Gizmos.color = new Color(1f, 0f, 0f, 0.4f);
